requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
crispy-bootstrap5==0.6
dj-database-url==0.5.0
Django==3.2.12
django-crispy-forms==1.14.0
django-currentuser==0.5.3
django-on-heroku==1.1.2
django-phonenumber-field==6.1.0
django-readonly-field==1.1.1
gunicorn==20.1.0
phonenumbers==8.12.44
Pillow==9.0.1
psycopg2-binary==2.9.3
pytz==2021.3
sqlparse==0.4.2
whitenoise==6.0.0
